pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "flick"
version = "1.0.0"
description = "A zero-boilerplate framework for building interactive ChatGPT widgets"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Flick Team", email = "hello@flick.dev"}
]
keywords = ["chatgpt", "widgets", "mcp", "framework", "react"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.20.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
flick = "flick.cli.main:cli"

[project.urls]
Homepage = "https://flick.dev"
Documentation = "https://flick.dev/docs"
Repository = "https://github.com/flick-framework/flick"
"Bug Reports" = "https://github.com/flick-framework/flick/issues"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["flick*"]
exclude = ["tests*"]

